pyproject.toml
[project]
name = "arbitrage-mvp"
version = "0.1.0"
description = "Solana arbitrage monitoring system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "solana>=0.36.10",
    "solders>=0.21.0",
    "pyyaml>=6.0.1",
    "loguru>=0.7.2",
    "websockets>=13.1",
    "httpx>=0.27.0",
    "pandas>=2.2.0",
    "numpy>=2.1.0",
]
